requirements.txt
pillow
imagehash
rich
questionary
flask
pywebview
send2trash
exifread
